app/requirements.txt
pandas==2.2.0

catboost==1.2.2

haystack-ai==2.0.0b7
qdrant-haystack==3.0.0

pydantic==2.6.1
pydantic_settings==2.2.1

PyMySQL==1.1.0
SQLAlchemy==2.0.25

streamlit==1.31.1
extra_streamlit_components==0.1.70
